Update crib-listing.component.html
[(ngModel)] is a 2 way data binding directive that binds an input , select , textarea (or custom form control) to a property on the scope. 
the 2way data binding [(ngModel)] which is sortfield is bound to the value, because sortfield is bound to our select list, within this we have options with values that are bound to them. so the sortfield is going to puck what ever option we select and passed the pipe argument.